packages/data_science/pyproject.toml
[project]
name = "data_science"
description = "Code location for dagster that exposes snowpark resources for ML-Ops"
version = "0.1.0"
readme = "README.md"
authors = [{name = "Andrew Staus"}]
requires-python = ">=3.13, <3.14"
dependencies = [
    "dagster-cloud==1.11.15",
    "dagster-snowflake==0.27.15",
    "dagster_postgres==0.27.15",
    "data-platform-utils",
    "faker==37.11.0",
    "psutil==7.1.0",
    "snowflake-ml-python>=1.19.0",
]

[dependency-groups]
dev = [
    "pytest>=8.4.2",
    "pytest-cov>=7.0.0",
    "pytest-xdist>=3.8.0",
]


[tool.uv.sources]
data-platform-utils = { path = "../../libs/data_platform_utils" }

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.dagster]
module_name = "data_science.definitions"
code_location_name = "data_science"

[tool.dg]
directory_type = "project"
project.root_module = "data_science"
